Working version

+++ src/client/Client.java
@@ -2,17 +2,17 @@ package client;
 
 import java.io.*;
 import java.net.*;
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.Map;
 import java.util.Scanner;
 
 import org.json.simple.JSONArray;
 import org.json.simple.JSONObject;
 
+import client.PrintHandler;
+
 /**
  * Tjadoop client
  *  upload big.mkv 100000000
+ *  130.229.155.38:1337
  */
 
 public class Client {
@@ -21,15 +21,18 @@ public class Client {
 		LS, RMDIR, MKDIR, DELETE, UPLOAD, DOWNLOAD
 	}
 
-	private static int connectionPort = 8989;
-	Socket socket = null;
-	private DataOutputStream out = null;
-	private DataInputStream in = null;
+	private static int masterPort = 1337;
+	public static int slavePort = 49384;
+	private static String slave1Address = "";
+
+	private static Socket socket = null;
+	public static DataOutputStream out = null;
+	public static DataInputStream in = null;
 	private final String master;
 
 	public Client(String address) throws NumberFormatException, Exception {
 		master = address;
-		//		openConnection(address);
+		//openConnection(address, masterPort);
 
 		Scanner scanner = new Scanner(System.in);
 		boolean doRun = true;
@@ -38,62 +41,59 @@ public class Client {
 			String input = scanner.nextLine();
 			String[] sArray = input.split(" ");
 
-			switch (sArray[0]) {
-			case ("ls"):
+			// TODO byt tillbaka till JRE 7 och använd switch, för i helskotta
+			//switch (sArray[0]) {
+			if (sArray[0].equals("ls")) {
 				System.out.println("Listing directory");
 				String dir = "";
 				if (sArray.length > 1)
 					dir = sArray[1];
 				listDirectory(dir);
-				break;
-			case ("rmdir"):
+			} else if (sArray[0].equals("rmdir")) {
 				if (!validateCommand(sArray)) {
 					System.out.println("Usage: rmdir <dir>");
 					break;
 				}
 				removeDirectory(sArray[1]);
 				System.out.println("Removing directory");
-				break;
-			case ("mkdir"):
+			} else if (sArray[0].equals("mkdir")) {
 				if (!validateCommand(sArray)) {
 					System.out.println("Usage: mkdir <dir>");
 					break;
 				}
 				makeDirectory(sArray[1]);
 				System.out.println("Creating directory");
-				break;
-			case ("upload"):
+			} else if (sArray[0].equals("up")) {
 				if (!validateCommand(sArray)) {
 					System.out.println("Usage: upload <dir> <duplication>");
 					break;
 				}
 				uploadFile(sArray);
-				break;
-			case ("download"):
+			} else if (sArray[0].equals("dl")) {
 				if (!validateCommand(sArray)) {
 					System.out.println("Usage: download <dir>");
 					break;
 				}
 				downloadFile(sArray);
 				System.out.println("Download file");
-				break;
-			case ("delete"):
+			} else if (sArray[0].equals("rm")) {
 				if (!validateCommand(sArray)) {
 					System.out.println("Usage: delete <dir/file>");
 					break;
 				}
 				deleteFile(sArray[1]);
 				System.out.println("Delete file");
-				break;
-			default:
+			} else
 				System.out.println("Not a valid command");
-			}
+			//			default:
 		}
+		//}
 		scanner.close();
 		socket.close();
 	}
 
-	private void openConnection(String address, int port) throws IOException {
+	public static void openConnection(String address, int port)
+			throws IOException {
 		closeConnection();
 
 		if (port == -1) {
@@ -114,7 +114,7 @@ public class Client {
 		}
 	}
 
-	private void closeConnection() throws IOException {
+	public static void closeConnection() throws IOException {
 		if (socket != null)
 			socket.close();
 		if (out != null)
@@ -150,16 +150,60 @@ public class Client {
 		} else
 			res = true;
 
-		if (input[0].equals("upload")) {
+		/*if (input[0].equals("upload")) {
 			if (input.length >= 3 && input[input.length - 1].matches("[0-9]+"))
 				res = true;
 			else {
 				System.out.println("Specify duplication amount");
 				return false;
 			}
-		}
+		}*/
 
 		return res;
+
+	}
+
+	private DataNodePartition[] getPartitionMapFromMaster(JSONArray resArray,
+			boolean isUpload) {
+		DataNodePartition[] dataNodePartitions = new DataNodePartition[resArray
+				.size()];
+
+		for (int i = 0; i < resArray.size(); ++i) {
+			JSONObject jsonObject = (JSONObject) resArray.get(i);
+			String ip = "";
+			if (i == 0 && isUpload) {
+				slave1Address = (String) jsonObject.get("ip");
+				ip = (String) jsonObject.get("ip");
+			} else {
+				ip = (String) jsonObject.get("ip");
+			}
+			JSONArray partitions = (JSONArray) jsonObject.get("partitions");
+			long[] starts = new long[partitions.size()];
+			long[] ends = new long[partitions.size()];
+
+			System.out.println("Partitions array json:");
+			System.out.println(partitions.toString());
+
+			for (int j = 0; j < partitions.size(); ++j) {
+				JSONObject partition = (JSONObject) partitions.get(j);
+				System.out.println("Json object: " + partition.toString());
+				starts[j] = (Long) partition.get("start");
+				ends[j] = (Long) partition.get("end");
+			}
+
+			System.out.println("starts");
+			for (int h = 0; h < starts.length; h++) {
+				System.out.println(starts[h]);
+			}
+			System.out.println("Ends");
+			for (int h = 0; h < ends.length; h++) {
+				System.out.println(ends[h]);
+			}
+
+			dataNodePartitions[i] = new DataNodePartition(starts, ends, ip);
+		}
+
+		return dataNodePartitions;
 
 	}
 
@@ -173,81 +217,72 @@ public class Client {
 	private File downloadFile(String[] input) throws ClassNotFoundException,
 			IOException {
 		String fileName = input[1];
-		//		Says what it wants to download to master. (file path)
-		//		Master can return "false" if it doesn't exists, obviously.
-		//		Otherwise, returns a list with data nodes List[id, bytesize].
-		//
-		//		Master redirects to nodes that contains the blocks to form the file.
-		//
-		//		Data nodes send blocks to client.
-		//
-		//		When all blocks have been received, client patches them together to a file.
-		//
-		//		Done!
 
 		/* Communication with master */
+
+		openConnection(master, masterPort);
 
 		MasterNodeCom.sendRequest(Request.DOWNLOAD, socket, out, fileName, 0);
 
 		JSONObject json = MasterNodeCom.getResponse(in);
 
-		boolean success = (boolean) json.get("ok");
+		boolean success = (Boolean) json.get("success");
 		if (!success) {
 			System.out
 					.println("Failed to download file, master does not approve");
 			return null;
 		}
 
-		String slaveAddr = "";
-		int slavePort = -1;
-		int hash = -1;
+		JSONArray resArray = (JSONArray) json.get("nodes");
+		System.out.println(resArray.toString());
 
-		slaveAddr = (String) json.get("ip");
-		slavePort = (int) json.get("port");
+		DataNodePartition[] dataNodePartitions = getPartitionMapFromMaster(
+				resArray, false);
 
-		/* Communication with slave */
+		//		String slaveAddr = (String) json.get("ip");
 
-		openConnection(slaveAddr, slavePort);
+		byte type = 1;
+		System.out.println("Downloading file " + fileName);
+		String filePath = "/" + fileName;
+		int fileHash = filePath.hashCode();
 
-		if (!DataNodeCom.download(in, fileName)) {
-			System.err.println("Failed to download.");
-			return null;
+		for (int p = 0; p < dataNodePartitions.length; p++) {
+			System.out.println("Geting partition from "
+					+ dataNodePartitions[p].ip);
+
+			//out.writeInt(starts.length);
+			//System.out.println("Num nodes: " + starts.length);
+
+			long[] starts = dataNodePartitions[p].starts;
+			long[] ends = dataNodePartitions[p].ends;
+
+			for (int k = 0; k < starts.length; k++) {
+
+				openConnection(dataNodePartitions[p].ip, slavePort);
+
+				System.out.println("Available inputstream?: " + in.available());
+
+				System.out.println("Sending creator data");
+				out.write(type);
+				out.writeInt(fileHash);
+
+				System.out.println("Type: " + type);
+				System.out.println("File hash: " + fileHash);
+
+				//out.writeUTF(dataNodePartitions[p].ip);
+				out.writeLong(starts[k]);
+				out.writeLong(ends[k]);
+
+				/* Communication with slave */
+				if (!DataNodeCom.download(in, fileName, dataNodePartitions[p],
+						starts[k], ends[k])) {
+					System.err.println("Failed to download.");
+					return null;
+				}
+			}
+
 		}
 
-		//Map<Integer, ByteTuple> blockMap = new HashMap<Integer, ByteTuple>();
-
-		//JSONArray jsonArray = (JSONArray) json.get("map");
-
-		//		for (int i = 0; i < jsonArray.size(); i++) {
-		//			JSONObject o = (JSONObject) jsonArray.get(i);
-		//			blockMap.put((Integer) o.get("key"),
-		//					new ByteTuple((long) o.get("start"), (long) o.get("end")));
-		//		}
-
-		/* Communication with slave */
-
-		//		openConnection(slaveAddr);
-
-		//		boolean notDone = true;
-		//		long downloadedBytes = 0;
-		//		int numberOfFullChunks = (int) Math.floor((filesize / chunksize));
-		//		int rest = (int) (filesize - (numberOfFullChunks * chunksize));
-		//		int i = 1;
-		//		while (notDone) {
-		//			Packet p;
-		//			if (i == numberOfFullChunks) {
-		//				p = Packet.readBytes(in, rest);
-		//				notDone = false;
-		//			} else
-		//				p = Packet.readBytes(in, chunksize);
-		//
-		//			byte[] bytes = p.getByteData();
-		//			//TODO gör något med bytsen, skriv till fil! Join från FileHandler
-		//			downloadedBytes += bytes.length;
-		//			i++;
-		//		}
-		//		System.out.println("Downloaded " + downloadedBytes + " bytes");
-		//
 		return null;
 	}
 
@@ -261,14 +296,19 @@ public class Client {
 	private boolean uploadFile(String[] input) throws IOException,
 			ClassNotFoundException {
 
+		System.out.println("Connecting to master");
+		openConnection(master, masterPort);
+		System.out.println("Connection successful");
+
 		System.out.println("Uploading file...");
 
 		String file = parseFilePath(input);
 
 		/* - Communication with master - */
-		String slaveAddr = "";
-		int slavePort = -1;
 		long filesize = new File(file).length();
+
+		//		System.out.println(filesize);
+		//System.out.println(out);
 
 		if (!MasterNodeCom.sendRequest(Request.UPLOAD, socket, out, file,
 				filesize))
@@ -277,21 +317,26 @@ public class Client {
 		/* Master response */
 		JSONObject json = MasterNodeCom.getResponse(in);
 
-		if (!(boolean) json.get("ok"))
+		if (!(Boolean) json.get("success")) {
+			System.out
+					.println("Failed to upload file, master does not approve");
 			return false;
+		}
 
-		slaveAddr = (String) json.get("ip");
-		slavePort = (int) json.get("port");
-		int[] ids = (int[]) json.get("idArray");
-		long[] starts = (long[]) json.get("startArray");
-		long[] ends = (long[]) json.get("endArray");
-		int hash = (int) json.get("hash");
+		//Gson gson = new Gson();
+		//String id = (String) json.get("id");
+		JSONArray resArray = (JSONArray) json.get("nodes");
+		System.out.println(resArray.toString());
+
+		DataNodePartition[] dataNodePartitions = getPartitionMapFromMaster(
+				resArray, true);
 
 		/* Communication with slave */
+		//closeConnection();
 
-		openConnection(slaveAddr, slavePort);
+		//System.out.println("Slave address: " + slave1Address);
 
-		if (!DataNodeCom.upload(out, file, filesize, ids, starts, ends, hash)) {
+		if (!DataNodeCom.upload(out, file, filesize, dataNodePartitions)) {
 			System.err.println("Failed to upload.");
 			return false;
 		}
@@ -307,7 +352,7 @@ public class Client {
 
 	private boolean makeDirectory(String dir) throws IOException,
 			ClassNotFoundException {
-		openConnection(master, connectionPort);
+		openConnection(master, masterPort);
 
 		/* form: /tobbe/skumpa , inte /tobbe/skumpa/ */// TODO fixa parser för detta
 
@@ -315,7 +360,7 @@ public class Client {
 			return false;
 		JSONObject json = MasterNodeCom.getResponse(in);
 
-		if (!(boolean) json.get("ok"))
+		if (!(Boolean) json.get("success"))
 			return false;
 
 		return true;
@@ -323,49 +368,56 @@ public class Client {
 
 	private boolean removeDirectory(String dir) throws IOException,
 			ClassNotFoundException {
-		openConnection(master, connectionPort);
+		openConnection(master, masterPort);
 
 		if (!MasterNodeCom.sendRequest(Request.RMDIR, socket, out, dir, 0))
 			return false;
 		JSONObject json = MasterNodeCom.getResponse(in);
 
-		if (!(boolean) json.get("ok"))
+		if (!(Boolean) json.get("success")) {
+			System.err.println("Failed to send to master");
 			return false;
+		}
 
+		// TODD vänta på en till ACK, för att directory har blivit borttaget
 		return true;
 	}
 
 	private String listDirectory(String dir) throws IOException,
 			ClassNotFoundException {
-		openConnection(master, connectionPort);
-		String ret = "";
-		//		new Packet(Packet.LS, dir).writeOut(out);
 
 		/* Master communication */
-		if (!MasterNodeCom.sendRequest(Request.LS, socket, out, dir, 0))
-			return null;
+
+		openConnection(master, masterPort);
+
+		MasterNodeCom.sendRequest(Request.LS, socket, out, dir, 0);
 
 		JSONObject json = MasterNodeCom.getResponse(in);
 
-		boolean ok = (boolean) json.get("ok");
-		if (!ok) {
+		if (!(Boolean) json.get("success")) {
 			System.out.println("Directory not found");
 			return null;
 		}
-		ret = (String) json.get("dir");
-		return ret;
+		JSONArray ret = (JSONArray) json.get("content");
+
+		System.out.println(ret.toString().replace("\\", ""));
+		return ret.toString().replace("\\", "");
 
 	}
 
 	private boolean deleteFile(String file) throws IOException {
-		openConnection(master, connectionPort);
+		openConnection(master, masterPort);
 
 		if (!MasterNodeCom.sendRequest(Request.DELETE, socket, out, file, 0))
 			return false;
 		JSONObject json = MasterNodeCom.getResponse(in);
 
-		if (!(boolean) json.get("ok"))
+		System.out.println(json.toString());
+
+		if (!json.get("cmd").equals("ack-rmfile")) {
+			System.err.println("Master does not approve");
 			return false;
+		}
 
 		return true;
 
@@ -381,6 +433,8 @@ public class Client {
 	 */
 	public static void main(String[] args) throws IOException,
 			ClassNotFoundException {
+
+		//		System.setOut(new PrintHandler("loggis.txt"));
 		try {
 			new Client(args[0]);
 		} catch (ArrayIndexOutOfBoundsException e) {
@@ -396,16 +450,6 @@ public class Client {
 			System.exit(1);
 		}
 
-	}
-
-	private class ByteTuple {
-		public long start;
-		public long end;
-
-		public ByteTuple(long s, long e) {
-			start = s;
-			end = e;
-		}
 	}
 
 }
